refactor: SerializedPlain
