+++ src/traits/impl_detail.rs
@@ -1,0 +1,3 @@
+mod serialized_plain;
+
+pub use serialized_plain::SerializedPlain;
