+++ src/traits/serde_encrypt_public_key.rs
@@ -1,4 +1,4 @@
-use core::{marker::PhantomData, ops::DerefMut};
+use core::ops::DerefMut;
 
 use crate::{
     error::Error,
@@ -12,6 +12,8 @@ use crypto_box::{
     ChaChaBox,
 };
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
+
+use super::SerializedPlain;
 
 /// Public-key authenticated encryption for serde-serializable types.
 ///
@@ -111,7 +113,7 @@ pub trait SerdeEncryptPublicKey {
     fn decrypt_ref<'de>(
         encrypted_message: &EncryptedMessage,
         combined_key: &ReceiverCombinedKey,
-    ) -> Result<ToDeserialize<Self>, Error>
+    ) -> Result<SerializedPlain<Self>, Error>
     where
         Self: Sized + Deserialize<'de>,
     {
@@ -127,7 +129,7 @@ pub trait SerdeEncryptPublicKey {
             .decrypt(nonce.into(), encrypted)
             .map_err(|_| Error::decryption_error("error on decryption of ChaChaBox"))?;
 
-        Ok(ToDeserialize::new(serial_plain))
+        Ok(SerializedPlain::new(serial_plain))
     }
 
     /// # Failures
@@ -139,39 +141,6 @@ pub trait SerdeEncryptPublicKey {
     {
         serde_cbor::to_vec(&self).map_err(|e| {
             Error::serialization_error(&format!("failed to serialize data by serde_cbor: {:?}", e))
-        })
-    }
-}
-
-/// Serialized plain-text.
-#[derive(Eq, PartialEq, Ord, PartialOrd, Hash, Debug)]
-pub struct ToDeserialize<T> {
-    serialized_plain: Vec<u8>,
-    _type: PhantomData<T>,
-}
-
-impl<T> ToDeserialize<T> {
-    fn new(serialized_plain: Vec<u8>) -> Self {
-        Self {
-            serialized_plain,
-            _type: PhantomData::default(),
-        }
-    }
-
-    /// Deserialize to get plain message.
-    ///
-    /// # Failures
-    ///
-    /// - [DeserializationError](crate::error::ErrorKind::DeserializationError) when failed to deserialize decrypted message.
-    pub fn deserialize<'de>(&'de self) -> Result<T, Error>
-    where
-        T: Sized + Deserialize<'de>,
-    {
-        serde_cbor::from_slice(&self.serialized_plain).map_err(|e| {
-            Error::deserialization_error(&format!(
-                "error on serde_cbor deserialization after decryption: {:?}",
-                e
-            ))
         })
     }
 }
